DBZ-19 Refactored the structure of the docker-images repository.
Reorganized the repository so that we can use DockerHub automated builds to build all versions of
our images any time those older versions are modified. For example, if we need to upgrade Java,
we can add a different folder for the new version, and change the Dockerfiles for previouly released
Debezium images to use that newer Java image; DockerHub should then automatically rebuild those
older versions.

Normally, we'll do this only when we need to upgrade versions of Java, Debian, or Kafka/Zookeeper.
When we release new versions of Debezium, we'll add corresponding folders and update the
automated builds to generate new tags for those new Debezium versions.