DBZ-1279 Rewrite `connect` Dockerfile
Only included versions 0.9 and 0.10 since they are the current ones
being built and the bases for the future ones.

The reasons behind it starts with:
./connect/0.10/Dockerfile:13 SC2039 In POSIX sh, brace expansion is undefined.
./connect/0.10/Dockerfile:13 SC2039 In POSIX sh, indirect expansion is undefined.
./connect/0.10/Dockerfile:13 SC2086 Double quote to prevent globbing and word splitting.
./connect/0.10/Dockerfile:13 DL4006 Set the SHELL option -o pipefail before RUN with a pipe in it

Now we have a simpler and future proof approach, where we do not define
complex shell structures within the RUN command, having a simple shell
scripts is a valida solution which is also easier to test.