DBZ-1279 Validate images upon build
This way we only validate those which are part of the current build,
also keeping compatibility with the Jenkins auto update by not adding
extra scripts with Debezium versions.
Another reason is to avoid to check the old versions.
The hadolint version is kept as `latest` on purpose, since it's a linter
we want to have access to its latest features and be able to update the
Dockerfiles timely for the upcoming backwards compatibility breaks