DBZ-3019 fix for hypen-containing environment variables
Please refer to: https://issues.redhat.com/browse/DBZ-3019

As requested in the issue commentary, a change for this script.

I imagine a bash guru could optimize the value extraction of this code into a single line.
I have changed the log4j creation code for completion, but this is likely unnecessary due to the constraints on package names.