DBZ-6258 Build also debezium/postgres:15 for ARM
It's really not clear to me what has caused DBZ-6258, but if we don't
build debezium/postgres:15 for ARM, it doesn't make any sense that
ARM build of image based on debezium/postgres:15 would succeed.

Build debezium/postgres:15 also for ARM. This seems to be the most
clean and easy solution.